meeting_minutes/requirements.txt
# Core dependencies for Meeting Minutes AI
crewai[tools]>=0.148.0,<0.153.0
streamlit>=1.47.0
openai>=1.97.0
python-dotenv>=1.1.0
pydub>=0.25.0
google-api-python-client>=2.177.0
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0

# ChromaDB with specific version for compatibility
chromadb>=0.4.0,<0.6.0

# File processing
pypdf>=5.8.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
python-docx>=1.1.0

# Audio processing
SpeechRecognition>=3.10.0

# Utilities
requests>=2.32.0
beautifulsoup4>=4.13.0
pandas>=2.3.0
numpy>=2.2.0
tqdm>=4.67.0
rich>=13.9.0
click>=8.2.0
typer>=0.16.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
